Report errors the rebar3 way
This commit also removes the catch-all clause when parsing CLI args,
which was used to error out because of the unrecognized arguments,
as it is impossible for this case to happen.

When rebar3 parses the CLI args, if it finds any that is not included
in the option list of the provider in use, it will error out by itself.